launch puppeteer function helper now uses plugin from browser-pool
